requirements.txt
fastapi==0.115.9
uvicorn==0.34.0
openai==1.65.1
python-dotenv==1.0.1
sqlalchemy==2.0.38
requests==2.32.3
psycopg2==2.9.10
httpx==0.28.1
langchain==0.3.20
langchain-openai==0.3.7
pydantic==2.10.6
pydantic-settings==2.6.0
google-generativeai==0.8.4
langsmith==0.3.11
langchain_community==0.3.18
langchain-core==0.3.44